backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.4.0
gunicorn==21.2.0
paramiko==3.4.0
mysqlclient==2.2.0

